==> Recommender Systems/04 - Exercices.ipynb <==
# Aula 05 - Avaliação de Sistemas de Recomendação - Exercícios

import pandas as pd
import numpy as np

### Importar base de dados

import wget
!python3 -m wget https://github.com/mmanzato/MBABigData/raw/master/ml-20m-compact.tar.gz
!tar -xvzf ml-20m-compact.tar.gz

movies = pd.read_csv('./dataset/movies_sample.csv')
ratings = pd.read_csv('./dataset/ratings_sample.csv')
df = ratings[['userId', 'movieId', 'rating']]
df = df.merge(movies[['movieId', 'title']])
df

movies_genres = movies.drop('genres', axis=1).join(movies.genres.str.split('|', expand=True)
             .stack().reset_index(drop=True, level=1).rename('genre'))
movies_genres.head()

movies_tags = pd.read_csv('./dataset/tags_sample.csv')
movies_tags.head()

map_users = {user: idx for idx, user in enumerate(df.userId.unique())}
map_items = {item: idx for idx, item in enumerate(df.movieId.unique())}

df['userId'] = df['userId'].map(map_users)
df['movieId'] = df['movieId'].map(map_items)

movies_tags['userId'] = movies_tags['userId'].map(map_users)
movies_tags['movieId'] = movies_tags['movieId'].map(map_items)
movies_tags.dropna(inplace=True)
movies_tags['movieId'] = movies_tags.movieId.astype(int)
movies_tags['userId'] = movies_tags.userId.astype(int)

movies_genres['movieId'] = movies_genres['movieId'].map(map_items)
movies_genres.dropna(inplace=True)
movies_genres['movieId'] = movies_genres.movieId.astype(int)

map_title = {}
for _, row in df.iterrows():
    map_title[row.movieId] = row.title


movies_genres[['movieId', 'genre']].to_csv('items_genres.dat', index=False, sep='\t', header=False)
movies_tags[['movieId', 'tag']].to_csv('items_tags.dat', index=False, sep='\t', header=False)

### Divisão da base em treino e teste

from sklearn.model_selection import train_test_split
train, test = train_test_split(df, test_size=.2, random_state=2)
train.to_csv('train.dat', index=False, header=False, sep='\t')
test.to_csv('test.dat', index=False, header=False, sep='\t')
train

### Exemplos de utilização do CaseRecommender

from caserec.recommenders.rating_prediction.item_attribute_knn import ItemAttributeKNN

ItemAttributeKNN('train.dat', 'test.dat', metadata_file='items_genres.dat', k_neighbors=10, as_similar_first=True).compute()
ItemAttributeKNN('train.dat', 'test.dat', metadata_file='items_tags.dat', k_neighbors=10, as_similar_first=True).compute()

***Exercício 01:*** Verifique o efeito no RSME ao aumentar o número de vizinhos do algoritmo ItemAttributeKNN com usando gêneros. Explique.

# TODO
k_neighbours = [10, 20, 50, 100, 200, 500]
for k in k_neighbours:
    model = ItemAttributeKNN('train.dat', 'test.dat', metadata_file='items_genres.dat', k_neighbors=k, as_similar_first=True,)
    model.compute( verbose=False, verbose_evaluation=False)
    print(f"K = {k}, RSME = {model.evaluation_results['RMSE']}")
    

O aumento no número de vizinhos *K* geralmente leva a uma melhoria na precisão das previsões, como indicado pela redução do RMSE. Porém, para a escolha ideal de *K* deve-se balancear entre ter um número suficiente de vizinhos para capturar padrões gerais e não ser tão grande a ponto de haver perdas de informações específicas.

***Exercício 02:*** Verifique o efeito na métrica MAP@1 ao aumentar o número de fatores do algoritmo BPRMF. Explique.

# TODO
from caserec.recommenders.item_recommendation.bprmf import BprMF
n_factors = [1, 4, 7, 10, 15, 20, 25, 30]
for n in n_factors:
    model = BprMF('train.dat', 'test.dat', 'ir_bprmf.dat', factors=n)
    model.compute( verbose=False, verbose_evaluation=False)
    print(f"Factors = {n}, MAP@1 = {model.evaluation_results['MAP@1']}")

O aumento no número de fatores geralmente melhora a métrica MAP@1 (*Mean Average Precision at 1*) , o que indica que o modelo BPRMF está se tornando mais eficaz em recomendar o filme relevante no topo da lista. Isso ocorre porque mais fatores permitem uma representação mais detalhada e precisa das preferências dos usuários e das características dos filmes. Como visto acima, ao se ter um aumento significante no números de fatores, é possível notar um aumento na métrica MAP@1.

***Exercício 03:*** Implemente a métrica Average Precision para um único usuário. Utilize as recomendações geradas pelo algoritmo BPRMF (consulte notebook Aula04_Exemplo como base).

model = BprMF('train.dat', 'test.dat', 'ir_bprmf.dat', factors=3)
model.compute(verbose=False, verbose_evaluation=False)

recommendations = pd.read_csv('ir_bprmf.dat', sep='\t', names=['userId', 'movieId', 'score'])

# TODO

# Calculate the average precision for a user
def average_precision(user_id, recommendations, test): 
    # Get the recommendations for the specific user
    recs_user = recommendations.loc[(recommendations.userId == user_id), 'movieId'].tolist()
    # Get the relevant movies for the specific user
    relevant_movies = test.loc[(test.userId == user_id), 'movieId'].tolist()
    
    n_relevant_movies = 0
    cumulative_precision = 0.0
    
    # Iterate over the recommendations and calculate the precision
    for i, movie in enumerate(recs_user):
        if movie in relevant_movies:
            n_relevant_movies += 1
            # Precision at index i
            precision_at_i = n_relevant_movies / (i + 1)
            cumulative_precision += precision_at_i
    
    if n_relevant_movies == 0:
        return 0.0
    
    ap = cumulative_precision / n_relevant_movies
    return ap

# Calc the average precision for the first 10 user_ids
for u in range(1, 11):
    print(f"User: {u}, AP = {average_precision(u, recommendations, test):.4f}") 

***Exercício 04:*** Um importante aspecto que pode ser avaliado em Sistemas de Recomendação é a diversidade da lista de recomendações. A métrica Intra-List Similarity (ILS) aplica uma função de similaridade (e.g. Cosseno, Jaccard, Pearson, etc.) entre todos os pares de itens da lista de recomendação, usando seus metadados como gêneros, tags, etc. Mais detalhes sobre essa métrica podem ser encontrados em: https://grouplens.org/site-content/uploads/Improving-WWW-20051.pdf

***a)*** Implemente uma função que calcula a ILS de uma lista de recomendação para um único usuário. Utilize os gêneros de filmes.

# TODO

***b)*** Utilize a função que implementou no item (a) para calcular a ILS de todos os usuários da base de dados. Utilize as recomendações geradas pelo algoritmo BPR MF do CaseRecommender.

# TODO

***c)*** Aplique a função que implementou no item (b) para calcular a ILS das recomendações geradas pelos algoritmos ItemKNN e ItemAttributeKNN. Compare os resultados de todos os algoritmos, verificando inclusive o que acontece com a acurácia (precisão, MAP ou NDCG) quando a diversidade aumenta ou diminui.

# TODO
