==> Recommender Systems/05 - Examples.ipynb <==
# Aula 06 - Filtragem Híbrida - Exemplos

## Criando dataframe do exemplo da aula

import pandas as pd
import numpy as np



data = [['Jessica', 'Oceans Eleven', 2], 
        ['Jessica', 'The Lion King', 4], 
        ['Jessica', 'Braveheart', 3], 
        ['Jessica', 'Independence Day', 2],
        ['Jessica', 'Deby&Loide', 3],
        ['Marta', 'The Princess Diary', 4], 
        ['Marta', 'Oceans Eleven', 3], 
        ['Marta', 'Braveheart', 4], 
        ['Marta', 'Independence Day', 3],
        ['Marta', 'Deby&Loide', 2],
        ['Jose', 'The Princess Diary', 1], 
        ['Jose', 'Oceans Eleven', 5], 
        ['Jose', 'The Lion King', 3], 
        ['Jose', 'Braveheart', 4], 
        ['Jose', 'Deby&Loide', 5],
        ['Dave', 'The Princess Diary', 1], 
        ['Dave', 'The Lion King', 2], 
        ['Dave', 'Braveheart', 3], 
        ['Dave', 'Independence Day', 4]]
        
genres = [['The Princess Diary', 'Drama|Comedia'],
          ['Oceans Eleven', 'Acao|Comedia'],
          ['The Lion King', 'Drama|Sci-Fi'],
          ['Braveheart', 'Drama|Acao'],
          ['Independence Day', 'Acao|Sci-Fi'],
          ['Deby&Loide', 'Comedia']]

df = pd.DataFrame(data, columns=['user','item', 'rating'])
movies = pd.DataFrame(genres, columns=['item','genres'])
df

movies

## Mapeando usuários e itens para ids

map_users = {user: idx for idx, user in enumerate(df.user.unique())}
map_items = {item: idx for idx, item in enumerate(df.item.unique())}
df['userId'] = df['user'].map(map_users)
df['itemId'] = df['item'].map(map_items)
df

movies['itemId'] = movies['item'].map(map_items)
movies

movies_genres = movies.drop('genres', axis=1).join(movies.genres.str.split('|', expand=True)
             .stack().reset_index(drop=True, level=1).rename('genre'))
movies_genres

### Funções para obter informações específicas do DataFrame

# Obter a nota que um usuário deu para um item.
def get_rating(userId,itemId):
    if len(df[(df['userId']==userId)&(df['itemId']==itemId)]) == 0:
        return 0
    return (df.loc[(df.userId==userId) & (df.itemId == itemId),'rating'].iloc[0])

get_rating(1, 5)

# Obter a lista de todos os itens que um usuário avaliou.
def get_item_ids(userId):
    if userId not in df['userId'].values:
        return []
    return (df.loc[(df.userId==userId),'itemId'].tolist())

get_item_ids(0)

def get_user_ids(itemId):
    if itemId not in df['itemId'].values:
        return []
    return (df.loc[(df.itemId==itemId),'userId'].tolist())

get_user_ids(2)

# Obter o título do item dado o seu id.
def get_item_title(itemId):
    if itemId not in df['itemId'].values:
        return ''
    return (df.loc[(df.itemId == itemId),'item'].iloc[0])

get_item_title(0)

# Obter a lista de ratings de um usuário.
def get_user_ratings(userId):
    if userId not in df['userId'].values:
        return []
    return (df.loc[(df.userId==userId),'rating'].tolist())

get_user_ratings(0)

# Obter a lista de ratings de um usuário.
def get_item_ratings(itemId):
    if itemId not in df['itemId'].values:
        return []
    return (df.loc[(df.itemId==itemId),'rating'].tolist())

get_item_ratings(0)

# Obter a média de ratings de um usuário
def get_user_mean(userId):
    return np.mean(get_user_ratings(userId))

get_user_mean(1)

# Obter a média de ratings de um item
def get_item_mean(itemId):
    return np.mean(get_item_ratings(itemId))

get_item_mean(1)

# Obter a lista de gêneros de um item
def get_genres(itemId):
    if itemId not in movies_genres['itemId'].values:
        return []
    return movies_genres.loc[(movies_genres.itemId==itemId),'genre'].tolist()

get_genres(0)

## Hibridização monolítica - combinação

### Computar a similaridade de itens usando gêneros e notas

from math import sqrt

def item_sim_pearson_rating(itemId1, itemId2):
    '''
    itemId1 & itemId2 : ids dos dois itens cuja similaridade será computada
    '''
    # Contar quantos usuários avaliaram os mesmos itens.
    item_list1 = get_user_ids(itemId1)
    item_list2 = get_user_ids(itemId2)
    common_users = list(set(item_list1) & set(item_list2))
    if len(common_users) == 0:
        return 0
    
    # Calcular a média de cada item
    item1_mean = get_item_mean(itemId1)
    item2_mean = get_item_mean(itemId2)
    
    # Cálculo da similaridade.
    sim = []
    norm1 = []
    norm2 = []
    for user in common_users:
        rating1 = get_rating(user, itemId1)
        rating2 = get_rating(user, itemId2)
        sim.append((rating1 - item1_mean)*(rating2 - item2_mean))
        norm1.append(pow(rating1 - item1_mean, 2))
        norm2.append(pow(rating2 - item2_mean, 2))
    
    return sum(sim) / (sqrt(sum(norm1)) * sqrt(sum(norm2)))

item_sim_pearson_rating(4, 5)

from math import pow, sqrt

def item_sim_jaccard_genre(itemId1, itemId2):
    '''
    itemId1 & itemId2 : ids dos dois itens cuja similaridade será computada
    '''
    # Similaridade baseada em gêneros (Jaccard).
    genre_list1 = get_genres(itemId1)
    genre_list2 = get_genres(itemId2)
    common_genres = list(set(genre_list1) & set(genre_list2))
    return len(common_genres) / len(set(genre_list1 + genre_list2))
        
item_sim_jaccard_genre(4, 5)

def hybrid_similarity_score(itemId1, itemId2, tradeoff):
    return tradeoff * item_sim_pearson_rating(itemId1, itemId2) + (1 - tradeoff) * item_sim_jaccard_genre(itemId1, itemId2)

hybrid_similarity_score(4, 5, 0.5)

### Obter os itens mais similares

def most_similar_items(itemId, k):
    '''
    itemId : Targeted item
    k : qtde de vizinhos
    '''
    # Obter lista de itens.
    item_ids = df.itemId.unique().tolist()
    
    # Obter a similaridade entre o item alvo e os demais itens
    sim = [(hybrid_similarity_score(itemId, i, 0.9), i) for i in item_ids if i != itemId]
    
    # Ordenação inversa.
    sim.sort()
    sim.reverse()
    
    # Retornando os usuários mais similares. 
    return sim[:k]

print(most_similar_items(0, 2))

### Calcular a nota

def get_prediction(userId, itemId, k):
    item_mean = get_item_mean(itemId)
    similar_items = most_similar_items(itemId, k)
    num = []
    den = []
    
    for s, j in similar_items:
        rj = get_rating(userId, j)
        if rj == 0:
            continue
        num.append(s * (rj - get_item_mean(j)))
        den.append(s)
    
    return item_mean + sum(num) / sum(den)

get_prediction(3, 4, 2)

## Hibridização monolítica - combinação - usando base de dados e CaseRecommender

### Importar base de dados

import wget
!python3 -m wget https://github.com/mmanzato/MBABigData/raw/master/ml-20m-compact.tar.gz
!tar -xvzf ml-20m-compact.tar.gz

movies = pd.read_csv('./dataset/movies_sample.csv')
ratings = pd.read_csv('./dataset/ratings_sample.csv')
df = ratings[['userId', 'movieId', 'rating']]
df = df.merge(movies[['movieId', 'title']])
map_users = {user: idx for idx, user in enumerate(df.userId.unique())}
map_items = {item: idx for idx, item in enumerate(df.movieId.unique())}
df['userId'] = df['userId'].map(map_users)
df['movieId'] = df['movieId'].map(map_items)
movies['movieId'] = movies['movieId'].map(map_items)
map_title = {}

for _, row in df.iterrows():
    map_title[row.movieId] = row.title
    
from sklearn.model_selection import train_test_split
train, test = train_test_split(df, test_size=.2, random_state=2)

### Calcular similaridade baseada em notas

from sklearn.metrics.pairwise import cosine_similarity
from scipy import sparse

n_users = train['userId'].max()
n_items = train['movieId'].max()

A = np.zeros((n_users+1,n_items+1))
for line in train.itertuples():
    A[line[1],line[2]] = line[3]

print("Original rating matrix shape: ",A.shape)
print("Matrix :\n",A)

from sklearn.metrics import pairwise_distances
from scipy.spatial.distance import cosine

#Valid metrics are ['euclidean', 'l2', 'l1', 'manhattan', 'cityblock', 'braycurtis', 'canberra', 'chebyshev', 'correlation', 'cosine', 'dice', 'hamming', 'jaccard', 'kulsinski', 'mahalanobis', 'matching', 'minkowski', 'rogerstanimoto', 'russellrao', 'seuclidean', 'sokalmichener', 'sokalsneath', 'sqeuclidean', 'yule', 'wminkowski', 'nan_euclidean', 'haversine'], or 'precomputed', or a callable
sim_matrix = pairwise_distances(A.T, metric="cosine")
print("Shape: ", sim_matrix.shape)
print(sim_matrix)

with open('sim_r_matrix.dat', 'w') as arq_sim_matrix:
    for i in range(len(sim_matrix)):
        for j in range(len(sim_matrix)):
            if i < j:
                arq_sim_matrix.write(str(i) + '\t' + str(j) + '\t' + str(sim_matrix[i][j]) + '\n')

### Testando só com similaridade baseada em ratings (equivalente ao ItemKNN)

from caserec.recommenders.rating_prediction.item_attribute_knn import ItemAttributeKNN

ItemAttributeKNN('train.dat', 'test.dat', similarity_file='sim_r_matrix.dat', as_similar_first=True).compute()

### Calcular similaridade baseada em gêneros

genres_unique = pd.DataFrame(movies.genres.str.split('|').tolist()).stack().unique()
genres_unique = pd.DataFrame(genres_unique, columns=['genre'])
print(genres_unique)
movies_exp = movies.join(movies.genres.str.get_dummies().astype(bool))
movies_exp.drop('genres', inplace=True, axis=1)
movies_exp.dropna(inplace=True)
B = movies_exp.iloc[:,2:].to_numpy(dtype=bool)

sim_g_matrix = pairwise_distances(B, metric="jaccard")
sim_g_matrix

with open('sim_g_matrix.dat', 'w') as arq_sim_matrix:
    for i in range(len(sim_g_matrix)):
        for j in range(len(sim_g_matrix)):
            if i < j:
                arq_sim_matrix.write(str(i) + '\t' + str(j) + '\t' + str(sim_g_matrix[i][j]) + '\n')

### Testando só com similaridade baseada em gêneros (equivalente a FBC)

ItemAttributeKNN('train.dat', 'test.dat', similarity_file='sim_g_matrix.dat', as_similar_first=True).compute()

### Hibridização

hib_rating_sim = pd.read_csv('sim_r_matrix.dat', sep='\t', names=['movieId1', 'movieId2', 'similarity'])
hib_rating_sim.sort_values(['movieId1', 'movieId2']).head()

hib_genre_sim = pd.read_csv('sim_g_matrix.dat', sep='\t', names=['movieId1', 'movieId2', 'similarity'])
hib_genre_sim.sort_values(['movieId1', 'movieId2']).head()

hib_sim = hib_rating_sim.merge(hib_genre_sim, on=['movieId1', 'movieId2'], how='outer')
hib_sim.head()

hib_sim[hib_sim['similarity_x'].isna()==True]

import math

def choose_sim(row):
    if math.isnan(row['similarity_x']):
        return row['similarity_y']        
    else:
        return 0.7*row['similarity_x'] + 0.3*row['similarity_y']

hib_sim['similarity'] = hib_sim.apply(lambda row: choose_sim(row), axis=1)
hib_sim

hib_sim[['movieId1', 'movieId2', 'similarity']].to_csv('sim_h_matrix.dat', index=False, header=False, sep='\t')

ItemAttributeKNN('train.dat', 'test.dat', similarity_file='sim_h_matrix.dat', as_similar_first=True).compute()

## Hibridização canalizada - meta-nível (MF + KNN)


from math import sqrt

def train_svdopt(train, n_factors, lr=0.05, reg=0.02, miter=10):
    global_mean = train['rating'].mean()
    n_users = df['userId'].max()+1
    n_items = df['movieId'].max()+1
    bu = np.zeros(n_users)
    bi = np.zeros(n_items)
    p = np.random.normal(0.1, 0.1, (n_users, n_factors))
    q = np.random.normal(0.1, 0.1, (n_items, n_factors))
    error = []
    for t in range(miter):
        sq_error = 0
        for index, row in train.iterrows():
            u = row['userId']
            i = row['movieId']
            r_ui = row['rating']
            pred = global_mean + bu[u] + bi[i] + np.dot(p[u], q[i])
            e_ui = r_ui - pred
            sq_error = sq_error + pow(e_ui, 2)
            bu[u] = bu[u] + lr * e_ui
            bi[i] = bi[i] + lr * e_ui
            for f in range(n_factors):
                temp_uf = p[u][f]
                p[u][f] = p[u][f] + lr * (e_ui * q[i][f] - reg * p[u][f])
                q[i][f] = q[i][f] + lr * (e_ui * temp_uf - reg * q[i][f])
        error.append(sqrt(sq_error/len(train)))
    
    return global_mean, bu, bi, p, q, error

gl, bu, bi, p, q, error = train_svdopt(train, 10, miter=30)

print('Items factors matrix shape: ', q.shape)

sim_f_matrix = pairwise_distances(q, metric="cosine")
print("Shape: ", sim_f_matrix.shape)
print(sim_f_matrix)

with open('sim_f_matrix.dat', 'w') as arq_sim_matrix:
    for i in range(len(sim_f_matrix)):
        for j in range(len(sim_f_matrix)):
            if i < j:
                arq_sim_matrix.write(str(i) + '\t' + str(j) + '\t' + str(sim_f_matrix[i][j]) + '\n')

ItemAttributeKNN('train.dat', 'test.dat', similarity_file='sim_f_matrix.dat', as_similar_first=True).compute()
