==> Recommender Systems/09 - Exercises.ipynb <==
# Aula 10 - Recomendação baseada em sessão - Exercícios

import pandas as pd
import numpy as np

### Leitura do arquivo 2019-Oct-sample.csv (vide Aula 10 - Exemplos) caso não possua o arquivo

subset = pd.read_csv('./2019-Oct-sample.csv')
subset.head()

map_items = {item: idx for idx, item in enumerate(subset.product_id.unique())}
map_sessions = {item: idx for idx, item in enumerate(subset.user_session.unique())}
subset['itemId'] = subset['product_id'].map(map_items)
subset['sessionId'] = subset['user_session'].map(map_sessions)
subset.head()

n_items = subset['itemId'].max()+1
print('No. items: ', n_items)
n_sessions = subset['sessionId'].max()+1
print('No. sessions: ', n_sessions)

# create a dataset
# remove sessions with less than 2 items
def create_data(df):
    df.sort_values(by=['sessionId', 'event_time'], inplace=True, ignore_index=True)
    sessions, session = [], []
    for index, value in df.iterrows():
        if index != 0:
            if value["sessionId"] == df.at[index-1, "sessionId"]:
                if value["event_type"] == 'view':
                    session.append(value["itemId"])
            else:
                if len(session) > 1:
                    sessions.append((df.at[index-1, "sessionId"], session))
                session = [value["itemId"]]
        else:
            session.append(value["itemId"])
    return sessions

sessions = create_data(subset)

import random

random.shuffle(sessions)
split = len(sessions) * 0.8
train = sessions[:int(split)]
test = sessions[int(split):]
print('No. train sessions: ', len(train))
print('No. test sessions: ', len(test))

def jaccard(list1, list2):
    intersection = len(list(set(list1).intersection(list2)))
    union = (len(list1) + len(list2)) - intersection
    return float(intersection) / union

actual_session = test[37]
target = actual_session[1][0:-1]
print(actual_session)
print(target)
subset.loc[subset.sessionId==actual_session[0]]

def compute_score(train, target, itemId):
    candidate_sessions = []
    for s in range(len(train)):
        if itemId in train[s][1]:
            candidate_sessions.append(train[s][1])
            
    score = 0
    for n in range(len(candidate_sessions)):
        score += jaccard(candidate_sessions[n], target)
    
    return score

categories = subset.loc[subset.sessionId==actual_session[0]]['category_code'].unique().tolist()
candidate_items = subset.loc[subset.category_code.isin(categories)]['itemId'].unique().tolist()
len(candidate_items)

ranking = []
for i in range(len(candidate_items)):
    ranking.append((compute_score(train, target, candidate_items[i]), candidate_items[i]))

ranking.sort()
ranking.reverse()
print(ranking[0:10])

subset.loc[subset.itemId==483]

***Exercício 01:*** A função compute_score() definida acima e explicada na aula, é a implementação do algoritmo Session-based KNN (S-KNN). Implemente uma variação da função que represente o algoritmo Sequential Session-based KNN (S-SKNN). Compare o desempenho de ambas as funções na recomendação do último item de uma sessão qualquer do conjunto de teste. Para fazer essa comparação, utilize a métrica Average Precision. 

# TODO

# Function that calculate the average precision for a session#
def average_precision(target_session, recommendations):
    # Get the recommended items
    recs_items = [rec[1] for rec in recommendations] 

    # Last item of the session (the target item that we want to recommend)
    relevant_item = target_session[-1]
    
    n_relevant_items = 0
    cumulative_precision = 0.0
    
    # Ierate over the recommendations and calc the precision
    for i, item in enumerate(recs_items):
        if item in relevant_item:
            n_relevant_items += 1
            # Precision at index i
            precision_at_i = n_relevant_items / (i + 1)
            cumulative_precision += precision_at_i
    
    # If no relevant item was found, return 0
    if n_relevant_items == 0:
        return 0.0
    
    # Calculate the average precision
    ap = cumulative_precision / n_relevant_items
    
    return ap


# Compute the recommendations using Session-based KMM (S-KNN)
ranking = []
for i in range(len(candidate_items)):
    ranking.append((compute_score(train, target, candidate_items[i]), candidate_items[i]))

ranking.sort()
ranking.reverse()

# Print the top 10 recommendations
print(ranking[0:10])

# Calculate and print the Average Precision for S-KNN
print(f"Average Precision S-KNN: {average_precision(actual_session, ranking)}")

# Function that calculate the score using Sequential Session-bases KNN (S-SKNN)
def compute_score_ssknn(train, target, itemId):
    candidate_sessions = []
    
    # Get all the sessions where the item appears
    for s in range(len(train)):
        if itemId in train[s][1]:
            candidate_sessions.append(train[s][1])
    
    score = 0
    
    # For each candidate session, calculate the weighted Jaccard similarity
    for n in range(len(candidate_sessions)):
        # Calculate the Jaccard similarity
        jaccard_sim = jaccard(candidate_sessions[n], target)
        
        weight = 0
        
        # Get the most recent item in target session that also appears in candidate session
        recent_common_item = None
        for i in range(len(target)-1, -1, -1):  
            if target[i] in candidate_sessions[n]:
                recent_common_item = target[i]
                break
        
        # If there is a recent commom item, calculate the weight
        if recent_common_item is not None:
            # Position of the recent item in target session
            target_position = target.index(recent_common_item) + 1
            
            # Calculate the weight
            weight = target_position / len(target) 

        # Multiply the Jaccard similarity by the weight
        weighted_jaccard = jaccard_sim * weight

        score += weighted_jaccard
    
    return score


# Compute the recommendations using Sequential Session-based KMM (S-SKNN)
ranking = []
for i in range(len(candidate_items)):
    ranking.append((compute_score_ssknn(train, target, candidate_items[i]), candidate_items[i]))

ranking.sort()
ranking.reverse()

# Print the top 10 recommendations
print(ranking[0:10])

# Calculate and print the Average Precision for S-SKNN
print(f"Average Precision S-SKNN: {average_precision(actual_session, ranking)}")

***Exercício 02:*** Implemente outra variação da função compute_score() que represente o algoritmo Sequential Filter Session-based KNN (SF-SKNN). Compare o desempenho desse algoritmo com as demais abordagens para uma sessão qualquer do conjunto de teste.

# TODO

# Function that calculate the score using Sequential Filter Session-bases KNN (SF-SKNN)
def compute_score_sfsknn(train, target, itemId):
    candidate_sessions = []
    
    # Pegar o último item do target
    # Get the last item of target session
    last_item_target = target[-1]
    
    # Get all sessions that contains the last item of the target session followed by the desired item (itemId)
    for s in range(len(train)):
        session = train[s][1]
        
        if last_item_target in session:
            last_item_index = session.index(last_item_target)
            
            # Check if the item appears right after the last item of target session
            if last_item_index + 1 < len(session) and session[last_item_index + 1] == itemId:
                candidate_sessions.append(session)
    
    score = 0
    
    # For each valid candidate session, calculate the Jaccard similarity
    for n in range(len(candidate_sessions)):
        score += jaccard(candidate_sessions[n], target)
    
    return score


# Compute the recommendations using Sequential Filter Session-based KMM (SF-SKNN)
ranking = []
for i in range(len(candidate_items)):
    ranking.append((compute_score_sfsknn(train, target, candidate_items[i]), candidate_items[i]))

ranking.sort()
ranking.reverse()

# Print the top 10 recommendations
print(ranking[0:10])

# Calculate and print the Average Precision for SF-SKNN
print(f"Average Precision SF-SKNN: {average_precision(actual_session, ranking)}")

***Exercício 03:*** Na aula utilizamos uma estratégia trivial para selecionar itens candidatos para poder calcular seu escore: selecionamos apenas itens da mesma categoria que os itens que estão na sessão atual. Isso pode ser um problema, pois numa sessão, um usuário pode estar visualizando um produto e o sistema poderia recomendar um produto de outra categoria (exemplo: usuário visualiza/compra um smartphone, e o sistema recomenda uma capa protetora). Pense e implemente uma estratégia para selecionar os itens candidatos para os quais será calculado o escore via função compute_score(). Lembre-se de que quanto menor a quantidade de itens candidatos, mais rápido o sistema irá gerar a recomendação top N. Explique sua estratégia.

# TODO

# The stategy consists in select candidate items by combining items 
# from the same category and items purchased in related sessions

# Get the categories of the items of current session
categories = subset.loc[subset.sessionId==actual_session[0]]['category_code'].unique().tolist()

# Get all the items from the current session
session_items = subset.loc[subset.sessionId == actual_session[0]]['itemId'].tolist()

# Find the sessions that contain these items
related_sessions = subset.loc[subset.itemId.isin(session_items)]['sessionId'].unique().tolist()

# Select rows that correspond to the related sessions
related_sessions_df = subset[subset.sessionId.isin(related_sessions)]

# Filter to get only the items that were purchased in the related sessions
purchased_items_df = related_sessions_df[related_sessions_df.event_type == 'purchase']

# Extract the unique purchased item IDs from those sessions
purchased_items = purchased_items_df['itemId'].unique().tolist()

# Get candidate items based on category (those in the same category as current session items)
candidate_items = subset.loc[subset.category_code.isin(categories), 'itemId'].unique().tolist()

# Combine candidate items with purchased items, removing duplicates
candidate_items = list(set(candidate_items + purchased_items))

print(len(candidate_items))

# Compute the recommendations using Sequential Filter Session-based KMM (SF-SKNN)
ranking_sknn = []
ranking_ssknn = []
ranking_sfsknn = []

for i in range(len(candidate_items)):
    ranking_sknn.append((compute_score(train, target, candidate_items[i]), candidate_items[i]))
    ranking_ssknn.append((compute_score_ssknn(train, target, candidate_items[i]), candidate_items[i]))
    ranking_sfsknn.append((compute_score_sfsknn(train, target, candidate_items[i]), candidate_items[i]))


ranking_sknn.sort()
ranking_sknn.reverse()

ranking_ssknn.sort()
ranking_ssknn.reverse()

ranking_sfsknn.sort()
ranking_sfsknn.reverse()

# Print the top 10 recommendations for each algorithm
print(f"SKNN Recommendations:\n{ranking_sknn[0:10]}")
print(f"\nS-SKNN Recommendations:\n{ranking_ssknn[0:10]}")
print(f"\nSF-SKNN Recommendations:\n{ranking_sfsknn[0:10]}\n")

# Calculate and print the Average Precision for each algorithm
print(f"Average Precision SKNN: {average_precision(actual_session, ranking_sknn)}")
print(f"Average Precision S-SKNN: {average_precision(actual_session, ranking_ssknn)}")
print(f"Average Precision SF-SKNN: {average_precision(actual_session, ranking_sfsknn)}")
