==> Recommender Systems/10 - Examples_Stanza.ipynb <==
!pip install stanza

import stanza
stanza.download('en') # download English model

nlp = stanza.Pipeline('en', processors='tokenize,mwt,pos,sentiment')

doc = nlp("Barack Obama was born in Hawaii. I am sad. Now, you are happy!")

doc.sentences[0].words

doc.sentences[0].words[0].xpos

if '': print(True)
